+++ hscontrol/db/users.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/juanfont/headscale/hscontrol/types"
@@ -61,9 +62,14 @@ func DestroyUser(tx *gorm.DB, uid types.UserID) error {
 	}
 
 	if len(nodes) > 0 {
+		blocking := make([]string, len(nodes))
+		for i, node := range nodes {
+			blocking[i] = fmt.Sprintf("%d (%s)", node.ID.Uint64(), node.Hostname)
+		}
+
 		return fmt.Errorf(
-			"%w: %d node(s) must be deleted or moved to another user first",
-			ErrUserStillHasNodes, len(nodes),
+			"%w: %d node(s) must be deleted or moved to another user first: %s",
+			ErrUserStillHasNodes, len(nodes), strings.Join(blocking, ", "),
 		)
 	}